Refactor: refactored component styles
